Rewrite to use Slack Bolt API
The previous bot was written as a Hubot app, but this rewrite is using
Slack's Bolt API, see https://slack.dev/bolt.

Per Slack, "Custom integrations will be deprecated and possibly removed
in the future." Their recommendation is to instead use a Slack app. It
ties this more tightly to Slack, but that's okay for now.